private-endpoints-with-appservice/without-private-vnet-demo/requirements.txt
psycopg>=3.1.0
psycopg-binary>=3.1.0
fastapi
uvicorn[standard]
azure-identity
